Move YamlPipelineJobConfiguration back to api module for extension (#21369)
* Move YamlPipelineJobConfiguration back to api module for extension

* Move YamlPipelineJobConfiguration back to api module for extension
